+++ Examples/Enzyme/Library/library.hpp
@@ -1,2 +1,2 @@
 #pragma once
-int square(int x);
+double square(double x);
